suspicious_trigger/requirements.txt
azure-functions
pandas
slack_sdk
azure-storage-blob
